backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
google-generativeai>=0.8.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
chromadb==0.4.18
PyPDF2==3.0.1
python-dotenv==1.0.0
httpx==0.25.1
pandas==2.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.1.0
langchain-community==0.0.10
docx2txt==0.8
pypdf==4.0.0
reportlab==4.0.8
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
